pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "r3dir"
version = "0.2.1.10"
authors = [
  {name="Horlad"},
]
description = "r3dir encoder/decoder package with CLI tool"
readme = "PYPIREADME.md" 
requires-python = ">=3.10"
dependencies = [
    "unishox2-py3",
    "pyperclip"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[tool.setuptools]
packages = ["r3dir"]

[tool.setuptools.package-data]
r3dir = ["hackvertor/encoder.js"]

[project.scripts]
r3dir = "r3dir._cli:_cli"

[project.urls]
"Homepage" = "https://github.com/Horlad/r3dir"
"Bug Tracker" = "https://github.com/Horlad/r3dir/issues"